[move] Move 'CurrencyUnit' Enum class
global 패키지 하위 enums 패키지로 이동

related to: #127
