+++ src/main/java/com/haejwo/tripcometrue/domain/store/dto/response/CityStoreResponseDto.java
@@ -1,6 +1,6 @@
 package com.haejwo.tripcometrue.domain.store.dto.response;
 import com.haejwo.tripcometrue.domain.city.entity.City;
-import com.haejwo.tripcometrue.domain.city.entity.CurrencyUnit;
+import com.haejwo.tripcometrue.global.enums.CurrencyUnit;
 import com.haejwo.tripcometrue.domain.store.entity.CityStore;
 import com.haejwo.tripcometrue.global.enums.Country;
