+++ src/test/java/com/haejwo/tripcometrue/domain/triprecord/repository/triprecord_schedule_image/TripRecordScheduleImageRepositoryTest.java
@@ -2,7 +2,7 @@ package com.haejwo.tripcometrue.domain.triprecord.repository.triprecord_schedule
 
 import com.haejwo.tripcometrue.config.TestQuerydslConfig;
 import com.haejwo.tripcometrue.domain.city.entity.City;
-import com.haejwo.tripcometrue.domain.city.entity.CurrencyUnit;
+import com.haejwo.tripcometrue.global.enums.CurrencyUnit;
 import com.haejwo.tripcometrue.domain.city.repository.CityRepository;
 import com.haejwo.tripcometrue.domain.member.entity.Member;
 import com.haejwo.tripcometrue.domain.member.repository.MemberRepository;
